+++ incubator/binding-pgsql.spec/src/main/scripts/io/aklivity/zilla/specs/binding/pgsql/streams/network/create.table.with.primary.key/server.rpt
@@ -53,7 +53,7 @@ write [0x53]                                 # type S
 
 write [0x53]                                 # type S
       35                                     # length
-      "standard_confirming_strings" [0x00]   # name
+      "standard_conforming_strings" [0x00]   # name
       [0x6f 0x6e 0x00]                       # value
 
 write [0x53]                                 # type S
