+++ incubator/binding-pgsql.spec/src/test/java/io/aklivity/zilla/specs/binding/pgsql/streams/ApplicationIT.java
@@ -12,7 +12,7 @@
  * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package io.aklivity.zilla.specs.binding.pgsql.streams.network;
+package io.aklivity.zilla.specs.binding.pgsql.streams;
 
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static org.junit.rules.RuleChain.outerRule;
@@ -26,10 +26,10 @@ import org.junit.rules.Timeout;
 import io.aklivity.k3po.runtime.junit.annotation.Specification;
 import io.aklivity.k3po.runtime.junit.rules.K3poRule;
 
-public class StreamIT
+public class ApplicationIT
 {
     private final K3poRule k3po = new K3poRule()
-        .addScriptRoot("net", "io/aklivity/zilla/specs/binding/pgsql/streams/network");
+        .addScriptRoot("app", "io/aklivity/zilla/specs/binding/pgsql/streams/application");
 
     private final TestRule timeout = new DisableOnDebug(new Timeout(10, SECONDS));
 
@@ -38,8 +38,8 @@ public class StreamIT
 
     @Test
     @Specification({
-        "${net}/create.table.with.primary.key/client",
-        "${net}/create.table.with.primary.key/server"
+        "${app}/create.table.with.primary.key/client",
+        "${app}/create.table.with.primary.key/server"
     })
     public void shouldCreateTableWithPrimaryKey() throws Exception
     {
@@ -48,18 +48,9 @@ public class StreamIT
 
     @Test
     @Specification({
-        "${net}/select.table/client",
-        "${net}/select.table/server" })
+        "${app}/select.table/client",
+        "${app}/select.table/server" })
     public void shouldSelectTable() throws Exception
-    {
-        k3po.finish();
-    }
-
-    @Test
-    @Specification({
-        "${net}/ssl.request/client",
-        "${net}/ssl.request/server" })
-    public void shouldHandleSslRequest() throws Exception
     {
         k3po.finish();
     }
